Extract constants and tags from StyleSheet file
